requirements.txt
pytorch
matplotlib
numpy
scipy 
sounddevice
soundfile
librosa
pandas
torchaudio
